Moved CompressedFile struct to separate file
